Home Task 9_2 (add)
Task 2
Maze Search Task
You have entered a maze. Somewhere in the maze is a magical lamp that will allow
you to return home. There is no other way out of the maze except by finding the
lamp. You (S) and lamp (D) can appear at different points in the maze.
Your task is to find the path (if it exists) to the lamp using the DFS and BFS methods.
As a result, output the correct path between S and D, the length of the path, and the
number of visited nodes. Obviously, you need to convert maze to graph.

Additional task
Graphical Representation of Maze
Draw the maze, source, destination, and path using more advanced graphics instead
of the console. You may use external libraries.